% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ForageR.R
\name{createForagers}
\alias{createForagers}
\title{Forager Constructor Wrapper}
\usage{
createForagers(
  numForagers,
  type = "Random",
  bounds = NA,
  locations = NA,
  bearings = NA,
  speeds = NA,
  persistences = NA,
  sights = NA,
  repeatAvoids = NA,
  quiet = FALSE,
  CRS = "NA",
  turnBiases = NA
)
}
\arguments{
\item{numForagers}{A numeric giving the number of foragers to be created}

\item{type}{A character giving the type of forager to create. Currenly handles "Random" (default, parent class Forager) and "BRW" (brwForager)}

\item{bounds}{A simple features collection of polygons within which all foragers must be created}

\item{locations}{A list of simple features collections each containing a single st_POINT giving the initial location of a new forager. If NA or empty, foragers will be created at random locations within bounds}

\item{bearings}{A single numeric giving the initial bearing of all new foragers in radians or a list of numerics equal in length to the value of numForagers}

\item{speeds}{A single numeric giving the initial speed of all new foragers or a list of numerics equal in length to the value of numForagers}

\item{persistences}{A single numeric giving the initial persistence of all new foragers or a list of numerics equal in length to the value of numForagers. Ignored unless type = "BRW"}

\item{sights}{A single numeric giving the initial sight range of all new foragers or a list of numerics equal in length to the value of numForagers}

\item{repeatAvoids}{A single numeric giving the number of unique patches a forager must visit before revisting a patch, or a list of numerics equal in length to the value of numForagers}

\item{quiet}{A logical indicating whether to suppress warnings about default values used to fill in empty fields}

\item{CRS}{A numeric giving the CRS code to assign to forager locations. If bounds argument is defined, the CRS of that object will be used and this argument will be ignored}

\item{turnBiases}{A single numeric giving the average turn angle of a forager in radians, or a list of numerics equal in length to the value of numForagers. Ignored unless type = "BRW"}
}
\description{
A wrapper function to create a list of foragers from lists of parameter values. Handles single values if parameter is the same for all foragers. Assigns default values for empty fields.
}
