% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ForageR.R
\docType{class}
\name{ddbrwForager-class}
\alias{ddbrwForager-class}
\alias{DDBRWForager}
\title{ddbrwForager constructor}
\description{
Constructor function for a Distance Discounting Biased Random Walk Forager class objects. Objects of this class have methods for finding targets, moving, and plotting. Movement switches between a biased random walk (with step lengths drawn from a gamma distribution and bearing drawn from a wrapped cauchy) and directed motion toward a target. Target selection is wieghted toward closer targets.
}
\section{Fields}{

\describe{
\item{\code{location}}{a simple features collection with a single POINT class object}

\item{\code{bearing}}{a numeric that gives the current bearing in radians. Default starting bearing is drawn from random uniform circular distribution}

\item{\code{speed}}{a numeric value that gives the scale parameter of the gamma distribution from which step lengths are draw. Because the shape parameter of this distribution is set to 1, the speed variable will equal the average step length}

\item{\code{sight}}{a numeric giving the distance at which the forager object can detect patches when in an environment}

\item{\code{path}}{a simple features collection with a single multipoint object containing previous locations of the forager. Usually objects are created with default multipoint objects containing only the initial location}

\item{\code{visitSeq}}{a character vector with the names (in order) of all patches visited. Must start with NAs equal in number to to the repeatAvoid variable (this is done by default if no value is given)}

\item{\code{targeting}}{a logical giving which mode of movement the forager is in. Initial value usually uses default "FALSE"}

\item{\code{repeatAvoid}}{numeric giving the number of different patches a forager must visit before targeting a recently visited patch again}

\item{\code{target}}{a simple feature data frame with a single row containing the target patch location and values}

\item{\code{turnBias}}{a numeric giving the radians from which the center of the wrapped cauchy distribution from which new bearings are drawn should be shifted from the previous bearing.}

\item{\code{persistence}}{a numeric between 0 and one which gives the concentration of the wrapped cauchy distribution from which new bearings are drawn, where a 1 results in a turning angle equal to turnBias every step and a 0 results in a circular random uniform probability distribution}
}}

\section{Methods}{

\describe{
\item{\code{tSelect(choices)}}{calculates probablilites and selects a target from a list of options but DOES NOT set the target (use setTarget for this, which calls tSelect). A forager will always select a patch it is on if possible. Otherwise, selection is random}
}}

