% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ForageR.R
\docType{class}
\name{distanceForager-class}
\alias{distanceForager-class}
\alias{dForager}
\title{dForager Constrctor}
\description{
Constructor function for distance Forager class objects. Obects of this class behave exactly like forager class objects, but when selecting a target, distance foragers prefer closer targets, discounting targets probability of selection by the cube of their distance
}
\section{Fields}{

\describe{
\item{\code{location}}{A simple features collection with a single POINT class object}

\item{\code{bearing}}{A numeric that gives the current bearing in radians. Default starting bearing is drawn from random uniform circular distribution}

\item{\code{speed}}{A numeric value that gives the scale parameter of the gamma distribution from which step lengths are draw. Because the shape parameter of this distribution is set to 1, the speed variable will equal the average step length}

\item{\code{sight}}{A numeric giving the distance at which the forager object can detect patches when in an environment}

\item{\code{path}}{A simple features collection with a single multipoint object containing previous locations of the forager. Usually objects are created with default multipoint objects containing only the initial location}

\item{\code{visitSeq}}{A character vector with the names (in order) of all patches visited. Must start with NAs equal in number to to the repeatAvoid variable (this is done by default if no value is given)}

\item{\code{targeting}}{A logical giving which mode of movement the forager is in. Initial value usually uses default "FALSE"}

\item{\code{repeatAvoid}}{Numeric giving the number of different patches a forager must visit before targeting a recently visited patch again}

\item{\code{target}}{A simple feature data frame with a single row containing the target patch location and values}
}}

\section{Methods}{

\describe{
\item{\code{tSelect(choices)}}{calculates probablilites and selects a target from a list of options but DOES NOT set the target (use setTarget for this, which calls tSelect). A forager will always select a patch it is on if possible. Otherwise, selection is random}
}}

