% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_path_data.R
\name{extract_path_data}
\alias{extract_path_data}
\title{Extract Path Data}
\usage{
extract_path_data(environment_list)
}
\arguments{
\item{environment_list}{a list with Environment class objects in each element}
}
\value{
list, paths data frame in first element, metadata in second
}
\description{
Extract Path Data
}
