% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ForageR.R
\name{generateBoundedPoint}
\alias{generateBoundedPoint}
\title{Bounded Point Generator}
\usage{
generateBoundedPoint(bounds)
}
\arguments{
\item{bounds}{A simple features collection containing polygons within which new point must be created.}
}
\description{
creates a spatial point object within the confines of a spatial feature (the input)
}
