% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ForageR.R
\name{getChoices}
\alias{getChoices}
\title{Choice Finder}
\usage{
getChoices(forager, patches)
}
\arguments{
\item{forager}{A single object of Forager class}

\item{patches}{A simple features data frame with a geom column containing the geometries of patches.}
}
\description{
Choice Finder
}
