% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ForageR.R
\name{logistic_growth}
\alias{logistic_growth}
\title{Logistic Growth}
\usage{
logistic_growth(y, max = 1, scale = 0.2, yadj = 0.5)
}
\arguments{
\item{y}{a vector of numerics representing current values experiencing logistic growth}

\item{max}{the horizontal asymptote of the growth function}

\item{scale}{the maximum growth rate}

\item{yadj}{scales the sigmoid midpoint on y axis, relative to max. At 0.5, maximum growth occurs at time 0}
}
\description{
A function to determine a patches location on its logistic growth curve (solved from VALUE), then increase VALUE by one unit of time on the curve
}
