% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_full_parameter_space.R
\name{run_full_parameter_space}
\alias{run_full_parameter_space}
\title{Simulate Over Parameter Space}
\usage{
run_full_parameter_space(
  num_patches = 2,
  forager_directedness = 0.5,
  forager_persistence = 0.6,
  forager_choice_determinism = 0,
  field_radius = 50,
  patch_radius = 1,
  patch_max_value = 10,
  patch_starting_value = 3,
  patch_regen = 0.03,
  forager_speed = 1,
  forager_sight = 20,
  forager_GUD = 0.3,
  forager_efficiency = 0.1,
  simulation_length = 2000
)
}
\arguments{
\item{num_patches}{numeric}

\item{forager_directedness}{numeric}

\item{forager_persistence}{numeric}

\item{forager_choice_determinism}{numeric}

\item{field_radius}{numeric}

\item{patch_radius}{numeric}

\item{patch_max_value}{numeric}

\item{patch_starting_value}{numeric}

\item{patch_regen}{numeric}

\item{forager_speed}{numeric}

\item{forager_sight}{numeric}

\item{forager_GUD}{numeric}

\item{forager_efficiency}{numeric}

\item{simulation_length}{numeric}
}
\value{
list containing objects of Environment Class
}
\description{
Allows complete parameterization of a set of simulations. Meant to be added to as more parameters are divised to be iterated over
}
