% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_iterators.R
\name{run_hypercube_sample}
\alias{run_hypercube_sample}
\title{Hypercube Sample Simulator}
\usage{
run_hypercube_sample(sample_value_df, num_steps = 2000, file_path = NA)
}
\arguments{
\item{sample_value_df}{a dataframe with named columns for each parameter in ForageR parameter space}

\item{num_steps}{the number of steps with to run each set of simulation parameters (rows in sample_value_df)}

\item{file_path}{character giving a path to save environments in prior to each simulation}
}
\value{
list of environment class objects
}
\description{
Hypercube Sample Simulator
}
