% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ForageR.R
\name{scale_attraction}
\alias{scale_attraction}
\title{Deterministic Attraction Modifier}
\usage{
scale_attraction(attractions, choice_determinism)
}
\arguments{
\item{attractions}{a vector of numerics representing attractiveness of various options}

\item{choice_determinism}{numeric. At 0, choice probability is directly proportional to attraction. When negative, the choice becomes more random. When positive, better options become relatively more attractive. Absolute values of 1 approach full randomness/determinism}
}
\description{
Modifies a vector of attraction values based on the deterministic nature of the choice. When choice determinism is positive, the best choice becomes relatively more attractive.
}
\examples{
\dontrun{
x <- seq(0,10,0.1)

plot(x, scale_attraction(x, 0))
lines(x, scale_attraction(x, 0.2), col = "green")
lines(x, scale_attraction(x, 0.5), col = "yellow")
lines(x, scale_attraction(x, 0.9), col = "red")
lines(x, scale_attraction(x, -0.2), col = "green")
lines(x, scale_attraction(x, -0.5), col = "yellow")
lines(x, scale_attraction(x, -0.9), col = "red")
}
}
