% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Foragers.R
\name{selector}
\alias{selector}
\title{selector}
\usage{
selector(choices)
}
\arguments{
\item{choices}{a datafame with a column named "prob". Values in this column should be numeric and add up to 1.}
}
\description{
selector
}
