% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ForageR.R
\name{seperate_patches}
\alias{seperate_patches}
\title{Patch Separator}
\usage{
seperate_patches(patch_geoms, iterations = 10)
}
\arguments{
\item{patch_geoms}{a spatial features collection (geometry list)#'}

\item{iterations}{#numeric giving number of interations to seperate before quitting with warning}
}
\value{
sfc with adjusted patch locations
}
\description{
checks for overlap of any geometries in patch_geoms, and shifts any overlapping patches away from each other. Will repeat up to "iterations" times
}
