% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ForageR.R
\name{to_sf_point}
\alias{to_sf_point}
\title{Make point sfc}
\usage{
to_sf_point(point, crs = NA)
}
\arguments{
\item{point}{A length two numeric vector, an st_POINT, or and sfc_POINT class object}

\item{crs}{A numeric giving the CRS code to apply to the point argument}
}
\description{
Make point sfc
}
